+++ src/main/java/com/social/instagram/config/AwsS3Config.java
@@ -22,6 +22,9 @@ public class AwsS3Config {
     @Value("${aws.s3.region}")
     private String region;
 
+    @Value("${aws.s3.bucket}")
+    private String bucket;
+
     @Bean
     @PostConstruct
     public AmazonS3 amazonS3Client() {
@@ -29,6 +32,11 @@ public class AwsS3Config {
                 .withCredentials(new AWSStaticCredentialsProvider(new BasicAWSCredentials(accessKey, secretKey)))
                 .withRegion(region)
                 .build();
+    }
+
+    @Bean
+    public String bucket() {
+        return bucket;
     }
 
 }
